requirements.txt
aiofiles==0.5.0
amqp==5.1.1
aniso8601==7.0.0
anyio==3.7.0
async-exit-stack==1.0.1
async-generator==1.10
beanie
billiard==4.1.0
celery
certifi==2022.12.7
chardet==4.0.0
click
dnspython==2.1.0
email-validator==1.1.2
exceptiongroup==1.1.1
fastapi==0.95.2
flower==1.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.3.1
lazy-model==0.0.5
MarkupSafe==2.0.1
motor==3.2.0
orjson==3.9.1
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
pydantic==1.10.9
pymongo
python-dateutil==2.8.2
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
ujson==4.0.2
urllib3==1.26.5
uvicorn
uvloop==0.17.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.6
websockets==10.0
